[+] Cisgender flag for #386
Okay I think if I added this, someone will ask about the straight flag again.

In terms of historical context, the cisgender flag has been used neutrally, and it has not been used as a symbol of dismissal like the straight flag did. While some people question whether it's necessary to have a pride flag for the majority identity that do not face discrimination, its existence isn't perceived as offensive by the LGBTQ+ community. So I judged it would be harmless to add the cisgender flag, while adding the straight flag would be harmful.